[Bridge][Twig] Decouple the SecurityExtension from its runtime
