+++ src/Symfony/Bridge/Twig/Extension/SecurityExtension.php
@@ -11,9 +11,6 @@
 
 namespace Symfony\Bridge\Twig\Extension;
 
-use Symfony\Component\Security\Acl\Voter\FieldVote;
-use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
-use Symfony\Component\Security\Core\Exception\AuthenticationCredentialsNotFoundException;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
 
@@ -24,37 +21,13 @@ use Twig\TwigFunction;
  */
 class SecurityExtension extends AbstractExtension
 {
-    private $securityChecker;
-
-    public function __construct(AuthorizationCheckerInterface $securityChecker = null)
-    {
-        $this->securityChecker = $securityChecker;
-    }
-
-    public function isGranted($role, $object = null, $field = null)
-    {
-        if (null === $this->securityChecker) {
-            return false;
-        }
-
-        if (null !== $field) {
-            $object = new FieldVote($object, $field);
-        }
-
-        try {
-            return $this->securityChecker->isGranted($role, $object);
-        } catch (AuthenticationCredentialsNotFoundException $e) {
-            return false;
-        }
-    }
-
     /**
      * {@inheritdoc}
      */
     public function getFunctions()
     {
         return array(
-            new TwigFunction('is_granted', array($this, 'isGranted')),
+            new TwigFunction('is_granted', array(SecurityRuntime::class, 'isGranted')),
         );
     }
